# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FDDT — Fraud Document Detection Tool API",
    "description": "FDDT (Fraud Document Detection Tool) backend. Environment: local.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/bulk-uploads": {
      "post": {
        "tags": [
          "bulk uploads"
        ],
        "summary": "Bulk upload: a zip of cases",
        "description": "Request body: the zip file itself (`Content-Type: application/zip`), not multipart. Query: `case_type` (applied to every case) and `filename` (the zip's name, for display). The zip holds one top-level folder per case, each containing that case's documents directly. The response (202) returns once the zip is stored: it lists every case folder found, with structural failures already marked (subfolder in a case folder, empty folder, empty or file over the company's per-file limit). Documents are then extracted, validated with the same checks as a single upload, and each valid case is created and queued, one case at a time. Poll GET /bulk-uploads/{id}. Partial success: a bad file is rejected on its own and a case with no valid file fails on its own. Zip rejections return `detail: {code, message, ...}` with `zip_empty`, `zip_too_large` (over the company's zip limit, `max_zip_size_mb`, 300 MB by default, decided from Content-Length before the body is read), `not_a_zip`, `zip_corrupted`, `zip_too_many_entries` or `zip_no_case_folders`.",
        "operationId": "create_bulk_upload_bulk_uploads_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CaseType",
              "description": "Case type for every case in the zip."
            },
            "description": "Case type for every case in the zip."
          },
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "description": "The zip's file name.",
              "title": "Filename"
            },
            "description": "The zip's file name."
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUploadDetail"
                }
              }
            }
          },
          "400": {
            "description": "`zip_empty`."
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Platform admins have read-only access."
          },
          "413": {
            "description": "`zip_too_large` (`size_bytes`, `max_size_bytes`)."
          },
          "415": {
            "description": "`not_a_zip`."
          },
          "422": {
            "description": "`zip_corrupted`, `zip_too_many_entries`, `zip_no_case_folders`."
          },
          "502": {
            "description": "Blob Storage rejected the upload."
          }
        }
      },
      "get": {
        "tags": [
          "bulk uploads"
        ],
        "summary": "List bulk uploads",
        "description": "Newest first. A `user` sees only their own; reviewers see their company's. Platform admins pass `company_id` (audited).",
        "operationId": "list_bulk_uploads_bulk_uploads_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Platform admins only (required for them): the company to act on. Company users are always confined to their own company.",
              "title": "Company Id"
            },
            "description": "Platform admins only (required for them): the company to act on. Company users are always confined to their own company."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BulkUploadSummary"
                  },
                  "title": "Response List Bulk Uploads Bulk Uploads Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bulk-uploads/{bulk_upload_id}": {
      "get": {
        "tags": [
          "bulk uploads"
        ],
        "summary": "Bulk upload summary (live per-case status)",
        "description": "Every case folder of the zip with its outcome (created / failed and why, per file) and, for created cases, live progress (queued → processing → done / flagged). Poll until `settled`. Another user's upload (for a `user`) or another company's is a 404.",
        "operationId": "get_bulk_upload_bulk_uploads__bulk_upload_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bulk_upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Bulk Upload Id"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Platform admins only (required for them): the company to act on. Company users are always confined to their own company.",
              "title": "Company Id"
            },
            "description": "Platform admins only (required for them): the company to act on. Company users are always confined to their own company."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUploadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkUploadDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "case_type": {
            "$ref": "#/components/schemas/CaseType"
          },
          "status": {
            "$ref": "#/components/schemas/BulkUploadStatus"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "zip_size_bytes": {
            "type": "integer",
            "title": "Zip Size Bytes"
          },
          "case_folder_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Folder Count"
          },
          "cases_created": {
            "type": "integer",
            "title": "Cases Created"
          },
          "cases_failed": {
            "type": "integer",
            "title": "Cases Failed"
          },
          "documents_accepted": {
            "type": "integer",
            "title": "Documents Accepted"
          },
          "documents_rejected": {
            "type": "integer",
            "title": "Documents Rejected"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "wrapper_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wrapper Folder"
          },
          "ignored_entries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ignored Entries",
            "default": []
          },
          "ignored_entry_count": {
            "type": "integer",
            "title": "Ignored Entry Count",
            "default": 0
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "progress": {
            "$ref": "#/components/schemas/BulkUploadProgress"
          },
          "settled": {
            "type": "boolean",
            "title": "Settled"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/BulkUploadCaseSchema"
            },
            "type": "array",
            "title": "Cases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "case_type",
          "status",
          "error_code",
          "error_message",
          "zip_size_bytes",
          "case_folder_count",
          "cases_created",
          "cases_failed",
          "documents_accepted",
          "documents_rejected",
          "uploaded_by",
          "created_at",
          "started_at",
          "finished_at",
          "progress",
          "settled",
          "cases"
        ],
        "title": "BulkUploadDetail"
      },
...
      "BulkUploadProgress": {
        "properties": {
          "validating": {
            "type": "integer",
            "title": "Validating",
            "default": 0
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "default": 0
          },
          "queued": {
            "type": "integer",
            "title": "Queued",
            "default": 0
          },
          "processing": {
            "type": "integer",
            "title": "Processing",
            "default": 0
          },
          "done": {
            "type": "integer",
            "title": "Done",
            "default": 0
          },
          "flagged": {
            "type": "integer",
            "title": "Flagged",
            "default": 0
          }
        },
        "type": "object",
        "title": "BulkUploadProgress"
      },
      "BulkUploadStatus": {
        "type": "string",
        "enum": [
          "queued",
          "ingesting",
          "complete",
          "failed"
        ],
        "title": "BulkUploadStatus"
      },
      "BulkUploadSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "case_type": {
            "$ref": "#/components/schemas/CaseType"
          },
          "status": {
            "$ref": "#/components/schemas/BulkUploadStatus"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "zip_size_bytes": {
            "type": "integer",
            "title": "Zip Size Bytes"
          },
          "case_folder_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Folder Count"
          },
          "cases_created": {
            "type": "integer",
            "title": "Cases Created"
          },
          "cases_failed": {
            "type": "integer",
            "title": "Cases Failed"
          },
          "documents_accepted": {
            "type": "integer",
            "title": "Documents Accepted"
          },
          "documents_rejected": {
            "type": "integer",
            "title": "Documents Rejected"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "case_type",
          "status",
          "error_code",
          "error_message",
          "zip_size_bytes",
          "case_folder_count",
          "cases_created",
          "cases_failed",
          "documents_accepted",
          "documents_rejected",
          "uploaded_by",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "BulkUploadSummary"
      },
...
      "CaseType": {
        "type": "string",
        "enum": [
          "school_document",
          "vendor_invoice",
          "commercial_invoice",
          "procurement_documentation",
          "quotation",
          "travel_reimbursement",
          "other"
        ],
        "title": "CaseType",
        "description": "Case-level category, selected at submission time (System Specification section\n3.6's document-type coverage list). This is separate from a document's\nown `document_type`, which is filled in later by the classification\nstep — a case can hold documents of more than one type (e.g. an\ninvoice plus its approval voucher)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name"
        ],
        "title": "UserSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}